Add implementation in Swift
The C macro is not usable in Swift, so I had to find a workaround. I
first used Objective-C methods that looked the same written in Swift as
the Macro does in Objective-C. This had one major drawback though: Test
results where printed in the implementation in `FBSnapshotTestCase.m`
not in the test itself.

All I did was converted the C macro to swift and used `XCTFail` with
`line` and `file` parameter. This is the same as it is used in Nimble.