add device agnostic tests & modify snapshot window fallback
Use UIApplication.keyWindow instead of a manually-created & installed
window when snapshotting a view w/o a window.  This prevents keyWindow
from being `nil` when generating device-agnostic reference image
filenames.

Accessing keyWindow is done through a category, so as to always throw an
assertion if it is nil, instructing developers to make sure their tests
have a host application w/ a key window.