Updated Readme.md to include adding import statement
When setting this up in my project, I followed the instructions and was still having trouble getting it to build.  I found the following answer which helped me out (http://stackoverflow.com/questions/27021323/can-you-use-the-fbsnapshottestcase-library-with-swift).  I believe anyone who imports the project would need to do this, if they are using swift, so let's just add it to the readme to help people get started.