adds a test for 'significantDeltaE'
This test is based on the comparison of two colors in
the CIE\*Lab colorspace.  It uses the simplest comparison,
aka CIE76, which is just a euclidean distance, but suitable
for simple comparison of two colors.  The difference
threshold of deltaE is 1.5, which is well below the accepted
value of 2.3 as 'just noticeable difference'